itemDetail 재작업 - 이미지 외 추가된 컬럼 데이터 표시 필요

+++ src/main/java/kr/co/anabada/item/entity/Bid.java
@@ -35,7 +35,7 @@ public class Bid {
 	@Column(nullable = false, columnDefinition = "ENUM('active','winning','lost','cancelled') default 'active'")
 	private String bidStatus;
 	@Column(nullable = false)
-	private Integer bidPrice;
+	private Long bidPrice;
 	@Column(nullable = false)
 	@UpdateTimestamp
 	private LocalDateTime bidTime;
